refactor: extract fuction
