+++ feed/utils.py
@@ -4,6 +4,7 @@ from dateutil import parser
 from lxml.etree import CDATA
 import lxml.etree as ET
 import gzip
+import re
 from gql import gql, Client
 from gql.transport.aiohttp import AIOHTTPTransport
 import logging
@@ -58,6 +59,24 @@ def tsConverter(s):
     timeorigin = parser.parse(s)
     timediff = timeorigin - datetime(1970, 1, 1, tzinfo=pytz.utc)
     return round(timediff.total_seconds() * 1000)
+
+def replace_alt_with_descrption(contentHtml, contentApiData, img_list):
+    for img in img_list:
+        if 'alt' in img:
+            alt = re.findall('alt=\".*?\"', img)
+            img_name = re.findall('\".*\"', alt[0])
+            img_name = img_name[0].replace('"', '')
+            for apidata_item in contentApiData:
+                if apidata_item['type'] == 'image' and 'name' in apidata_item['content'][0] and apidata_item['content'][0]['name'] == img_name:
+                    if 'title' in apidata_item['content'][0]:
+                        new_description = apidata_item['content'][0]['title']
+                    else:
+                        new_description = ''
+                    contentHtml = contentHtml.replace(
+                        alt[0], f'alt="{new_description}"')
+                    break
+    return contentHtml
+
 
 def sub(parentItem, tag, content=None):
     element = ET.SubElement(parentItem, tag)
